add getValues and interpolate
